ClientChannel commands state, ChatLive reads from new interfaces
Now that ChatSession and ChatSessionRegistry don't know about game start events, ClientChannel controls the logic for looking up and starting chat sessions rather than acting as a passthrough to ChatSessionRegistry. This provides a better separation of concerns, as now the registry just cares about chat sessions, and the client channel serves as a source of truth for whether a client is connected and how that client's actions translate to interactions with the chat application.